requirements.txt
streamlit
pandas
numpy
gspread
requests
pillow
